# src/main/java/com/match/matchodds/dto/MatchDto.java
package com.match.matchodds.dto;

import com.match.matchodds.model.Sport;
import lombok.Data;

import java.time.LocalDate;
import java.time.LocalTime;
import java.util.List;

@Data
public class MatchDto {
    private Long id;
    private String description;
    private LocalDate matchDate;
    private LocalTime matchTime;
    private String teamA;
    private String teamB;
    private Sport sport;
    private List<MatchOddsDto> matchOdds;
}